requirements.txt
sentence-transformers
openai
chromadb
python-dotenv
